[opengl] Split gles backend and enable tests
This PR is the minimal changeset so that we can enable tests for the
newly added gles backend.

It does a few things:
- Remove `use_gles` in `ti.init()` and `CompileConfig`, it's now just a
new backend `ti.gles`.
- `ti.reset()` now destructs glfw resources created in the global scope.
Note in the future we'd like these to be managed by `Program` but
that'll make this PR too large to review.
- Properly enable testing for gles backend and skip when necessary. Most of them are just
the same as ones we skipped for opengl backend.

There're also a few followups that I'll send after this PR since we'd
want those to be tested against both Opengl and Gles backends:
- split gles and opengl initialization
- split opengl initialization and destruction
- maybe also cleanup other backends

ghstack-source-id: dcf39bc66415ea9d5d537c0abde359dcd7126613
Pull Request resolved: https://github.com/taichi-dev/taichi/pull/6198